Revamp main menu #2: Electric Boogaloo (#30)
* Clean up button CSS, add and update buttons

Optimizes the button CSS, adds the editor button, updates the Discord button to the SMAS style and updates the leaderboard button (thanks Captain)

* Add max length of 20 to name input fields

* Remove stray white pixel from marselect outline

* Tiny control menu optimization

* Revamp changelog; goodbye patch notes page

The changelog is now its own popup menu on the main menu. The patch notes page has been removed and replaced with a Markdown document, patch.md. Changelogs button has been renamed to Changelog. Removed changelogs as popup window from to-do list.

* Fix changelog not showing if logged in

* 404 if profile is not found instead of 200

* Update settings menu

• Redesign audio sliders (shorter white slider with a mushroom thumb)
• Add audio slider numbers
• Replace X with *

* Fallback to bkga on thin screens

Hides the main canvas in favor of the bkga if the rendering surface is too thin. Despite being deprecated, device-width is used instead of width so it only triggers on mobile devices, not thin browser windows.

* Darken note color, remove text shadow

* Add 3.1.0 patch notes in a branch called 3.0.0 wow

---------

Co-authored-by: terminalarch <terminalartix@mail.com>